+++ platform/ios/scripts/style-spec.js
@@ -1,1 +1,1 @@
-module.exports = require('../../../mapbox-gl-js/src/style-spec/reference/v8');
+module.exports = require('../../../maplibre-gl-js/src/style-spec/reference/v8');